Move explore index page to app router
